Add Jackson module to serialize/deserialize FieldValue
This commit adds a new `GraphQlModule` Jackson module that supports:

* serialization/deserialization of `FieldValue<T>`
* the upgrade of `FieldValue<T>` as reference types

This enables `FieldValue<T>` to be read/written from/to JSON with
Jackson on the client side. Note, on the server side this module is not
involved as the payload is deserialized as a `Map<String,Object>` and
values are bound as `FieldValue<T>` types directly.

Closes gh-1174

Signed-off-by: James Bodkin <james.bodkin@amphora.net>
[brian.clozel@broadcom.com: apply code conventions, use FieldValue type]
Signed-off-by: Brian Clozel <brian.clozel@broadcom.com>