Fix high memory usage in ContextPropagationHelper
Prior to this commit, gh-1149 added support for cancellation detection
at the transport level and the propagation of the CANCEL signal to data
fetcher `Publisher`. `ContextPropagationHelper` stores in the context a
`Sink` that emits the cancel signal. All upstream publishers are then
subscribing to this and canceling themselves if they receive a signal.

In the case of high field count queries + high request/sec services,
this would create a significant memory overhead.

This commit reduces the support to the essential parts:
* avoid further calls to data fetchers if the request is canceled
* cancel subscription publishers

This means other in-flight data fetchers that are publisher-based are
not canceled anymore. This should be a good compromise for high RPS

See gh-1242