Support option input values in client requests
This commit adds support for optional input values in GraphQL client
requests, using `ArgumentValue<T>`.
This commit adds new Jackson 3.x and 2.x modules that support the
serialisation of `ArgumentValue<T>` in input types. Note that this is
not meant to be used on the server side, and deserialization is not
supported.

Closes gh-1264