Throw client exception when failing to decode field
Prior to this commit, calling `ClientResponseField.toEntity` on a valid
field could throw a `DecodingException` if the requested type for
deserialization is not compatible with the serialized value.

When thrown in a server application, this exception is handled as an
HTTP "Bad Request". This hides the server error.

This commit catches all exceptions thrown during the decoding process
and wraps with in `GraphQlClientException` to avoid such behavior.

Closes gh-1219