Fix missing response body in client exceptions
As of gh-1117, the `HttpSyncGraphQlClient` supports the "GraphQL over
HTTP" specification. More specifically, it now handles HTTP 4xx
responses sent by GraphQL servers when the response content type is
"applcation/graphql-response+json".

This change introduced a regression where 4xx and 5xx HTTP responses now
always throw `HttpClientErrorException` (instead of throwing
`HttpClientErrorException` or `HttpServerErrorException` depending on
the case), and this exception is missing the response body and other
information from the response.

This commit ensures that exceptions are thrown in a similar fashion to
the default `StatusHandler` from Framwork.

Fixes gh-1259