Avoid further setter binding when possible
As of gh-1163, the `GraphQlArgumentBinder` can bind using both a
constructor and setters. This is especially useful for Kotlin data
classes. This change introduced a significant performance penalty for
large and deep data sets.

This commit ensures that we are only performing setter binding for input
data that was not already consumed during the constructor binding phase.

Fixes gh-1284