Add schema inspection support for nullness checks
Prior to this commit, the schema inspection feature would detect missing
fields and arguments when looking at class properties and registered
data fetchers.

The GraphQL schema also scpecifies the nullability of fields and
arguments, using the "!" marker type.

This commit enhances the schema inspector to read the nullability
information from the schema and compare it to the nullability
information present in the code. This is only active if the application
code is written in Kotlin or is using JSpecify annotations.
If a nullability mismatch is detected (NULL vs NON_NULL), the
information is displayed in the inspection report.

Note that by default, Java types have the "UNSPECIFIED" nullness
information, because we cannot infer the nullability from typical Java
classes.

This feature is currently only supported if the `DataFetcher` is
actually backed by a Java method, a Kotlin function or a class property.
In other cases, the type use information will not be available.

Closes gh-1220