Fix high cardinality issue with "graphql.operation"
Prior to this commit, the GraphQL observability instrumentation would
add a "graphql.operation" low cardinality keyvalue to the request
observations. This value is extracted from the `ExecutionInput`.
But this value can be effectively high cardinality as it is driven by
user input and clients can set any name there.

This commit fixes this cardinality issue by:

* renaming "graphql.operation" to "graphql.operation.name" and making it
  high cardinality.
* adding a new "graphql.operation.type" low cardinality keyvalue that is
  extracted from the `ExecutionContext`, after the request document has
  been parses and validated. This information is now used for the trace
  context name.

Fixes gh-1322