Fix NullPointerException in SchemaMapping inspector
Prior to this commit, the new nullness checks in the schema mapping
inspection (see gh-1220) would fail for null field definitions. While
not common in schemas, such types can be cnotributed by the Federation
support for the JVM.

This commit ensures that such types are considered as nullness
"unspecified" and does not break the inspection process.

Fixes gh-1333