Skip nullness check for Map-like return types of @BatchMapping methods
Prior to this commit, the schema inspector would perform nullness checks
for fields and data fetchers. `@BatchMapping` methods can return
`Map<Book, Author>` types and such cases can be raised as inconsistent
with the schema declaration. This check is invalid because we cannot
infer the nullness of the actual schema field given a `Map` fetched
using a `DataLoader`.

This commit ensures that such cases are skipped.

Closes gh-1334