Refine nullness checks in the schema report inspection
Prior to this commit, gh-1220 introduced nullness checks in the schema
mapping inspection process. This generally reports all mismatches
between the GraphQL schema nullness information and the corresponding
Java type nullness.

This commit refines the implementation with the following changes:

If an application type is "non-null" and the GraphQL schema type is
"nullable", we should not report that as a problem. Partial responses
is a key GraphQL feature and we should not force developers to make
fields "non-null" just because applications types are "non-null".
If a data fetching error happens, the engine will null out fields in the
hierarchy up until null is a valid value.

The method names and report information now use "errors" instead of
"mismatches" as a consequence.

Closes gh-1344