Further refine nullness checks in schema inspection
This commit further refines nullness checks in the schema mapping
inspection to process argument errors differently: unlike fields, where
the application should not return `null` when the schema does not allow
it, arguments take user input. This means we should only treat as errors
if the schema accept nullable fields but the applicaiton code does not.

Closes gh-1344